rename resources -> model
