+++ crates/dessins/src/model.rs
